[Silver III] Title: 주유소, Time: 388 ms, Memory: 46676 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -3,35 +3,35 @@ import java.util.*;
 
 public class Main {
     static int n;
-    static int[] distList;
-    static int[] priceList;
+    static Long[] distList;
+    static Long[] priceList;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 
         n = Integer.parseInt(br.readLine());
-        distList = new int[n-1];
-        priceList = new int[n];
+        distList = new Long[n-1];
+        priceList = new Long[n];
 
         StringTokenizer st = new StringTokenizer(br.readLine());
         for(int i=0; i<n-1; i++){
-            distList[i] = Integer.parseInt(st.nextToken());
+            distList[i] = Long.parseLong(st.nextToken());
         }
         st = new StringTokenizer(br.readLine());
         for(int i=0; i<n; i++){
-            priceList[i] = Integer.parseInt(st.nextToken());
+            priceList[i] = Long.parseLong(st.nextToken());
         }
 
-        int sum = 0;
-        int minPrice = priceList[0];
+        Long sum = 0L;
+        Long minPrice = priceList[0];
         for(int i=1; i<n; i++){
             if(priceList[i] > minPrice){
-                int num = distList[i-1] * minPrice;
+                Long num = distList[i-1] * minPrice;
                 sum += num;
             }
             else if(priceList[i] <= minPrice){
-                int num = minPrice * distList[i-1];
+                Long num = minPrice * distList[i-1];
                 sum += num;
 
                 minPrice = Math.min(minPrice, priceList[i]);
